+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changes by Version
 
 <!-- next version -->
+## v1.24.0
+- Upgrade `logzio_client_terraform` to `1.24.1`.
+- Support Consumption plan in `sub_accounts`.
+  - Add `SoftLimitGB`
+- Mark `sub_accounts.flexible` field as `Computed` (since it cannot be edited).
+
 ## v1.23.0
 - Upgrade `logzio_client_terraform` to `1.27.4`.
 - Metrics Drop Filters: Add optional `drop_policy` field with values `DROP_BEFORE_PROCESSING` (default) and `DROP_BEFORE_STORING`.